requirements.txt
Django==1.9.6
django-filter==0.13.0
djangorestframework==3.3.3
Markdown==2.6.6
psycopg2==2.6.1
